+++ lib/common/ForwardDecl.h
@@ -23,7 +23,7 @@ class ThreadPool;
 template <typename T>
 class ThreadLocal;
 
-template <typename T>
+template <typename T, typename TAllocator>
 class List;
 
 template <typename T>