+++ lib/physics/Eos.h
@@ -179,5 +179,24 @@ public:
     }
 };
 
+/// \brief ANEOS equation defined by a look-up table.
+class Aneos : public IEos {
+private:
+public:
+    explicit Aneos(const BodySettings& settings);
+
+    virtual Pair<Float> evaluate(const Float rho, const Float u) const override;
+
+    /// Currently not implemented.
+    virtual Float getInternalEnergy(const Float UNUSED(rho), const Float UNUSED(p)) const override {
+        NOT_IMPLEMENTED;
+    }
+
+    /// Currently not implemented.
+    virtual Float getDensity(const Float UNUSED(p), const Float UNUSED(u)) const override {
+        NOT_IMPLEMENTED;
+    }
+};
+
 
 NAMESPACE_SPH_END